Simplify GoogleBook model structure into one file
