--- app/src/main/java/com/readtracker/android/googlebooks/model/ImageLinks.java
@@ -1,10 +1,0 @@
-package com.readtracker.android.googlebooks.model;
-
-import com.google.gson.annotations.SerializedName;
-
-public class ImageLinks {
-  @SerializedName("thumbnail") String thumbNail;
-  public String getThumbNail() {
-    return thumbNail;
-  }
-}
