+++ app/src/main/java/com/readtracker/android/support/SimpleAnimationListener.java
@@ -3,15 +3,7 @@ package com.readtracker.android.support;
 import android.view.animation.Animation;
 
 public abstract class SimpleAnimationListener implements Animation.AnimationListener {
-  @Override public void onAnimationStart(Animation animation) {
-
-  }
-
-  @Override public void onAnimationEnd(Animation animation) {
-
-  }
-
-  @Override public void onAnimationRepeat(Animation animation) {
-
-  }
+  @Override public void onAnimationStart(Animation animation) {}
+  @Override public void onAnimationEnd(Animation animation) {}
+  @Override public void onAnimationRepeat(Animation animation) {}
 }
